Move user.entity to parent folder
