+++ src/orders/entities/order.entity.ts
@@ -1,5 +1,5 @@
 import { Entity, Column, PrimaryGeneratedColumn, BaseEntity, ManyToOne, OneToMany } from "typeorm";
-import { User } from "../../users/entities/user.entity";
+import { User } from "../../users/user.entity";
 import { OrderRow } from "./order-row.entity";
