+++ src/ratings/rating.entity.ts
@@ -1,6 +1,6 @@
 import { Entity, Column, PrimaryGeneratedColumn, BaseEntity, ManyToOne } from "typeorm";
 import { Product } from "../products/product.entity";
-import { User } from "../users/entities/user.entity";
+import { User } from "../users/user.entity";
 
 @Entity()
 export class Rating extends BaseEntity {
