+++ src/users/users.controller.ts
@@ -1,6 +1,6 @@
 import { Body, Controller, Delete, Get, Param, Patch, Post, Request, UseGuards } from "@nestjs/common";
 import { UsersService } from "./users.service";
-import { User } from "./entities/user.entity";
+import { User } from "./user.entity";
 import { CreateUserDto } from "./dto/create-user.dto";
 import { UpdateUserDto } from "./dto/update-user.dto";
 import { Roles } from "../auth/roles/roles.decorator";
